+++ src/zenml/zen_server/template_execution/utils.py
@@ -462,123 +462,3 @@ def deployment_request_from_template(
     )
 
     return deployment_request
-
-
-from enum import Enum
-from typing import Any, Dict, Optional
-
-from pydantic import create_model
-
-from zenml.client import Client
-from zenml.config import ResourceSettings
-from zenml.config.step_configurations import StepConfigurationUpdate
-from zenml.stack import Flavor
-
-
-def generate_config_schema(
-    stack: StackResponse, config: PipelineRunConfiguration
-) -> Dict[str, Any]:
-    """Generate a JSON config schema for the stack and run configuration.
-
-    Args:
-        stack: The stack.
-        config: The run configuration.
-
-    Returns:
-        The generated schema dictionary.
-    """
-    experiment_trackers = []
-    step_operators = []
-
-    settings_fields = {"resources": (ResourceSettings, None)}
-    for component_list in stack.components.values():
-        assert len(component_list) == 1
-        component = component_list[0]
-        flavors = Client().zen_store.list_flavors(
-            FlavorFilter(name=component.flavor, type=component.type)
-        )
-        assert len(flavors) == 1
-        flavor_model = flavors[0]
-        flavor = Flavor.from_model(flavor_model)
-
-        for class_ in flavor.config_class.__mro__[1:]:
-            # Ugly hack to get the settings class of a flavor without having
-            # the integration installed. This is based on the convention that
-            # the static config of a stack component should always inherit
-            # from the dynamic settings.
-            if issubclass(class_, BaseSettings):
-                if len(class_.model_fields) > 0:
-                    settings_key = f"{component.type}.{component.flavor}"
-                    settings_fields[settings_key] = (
-                        class_,
-                        None,
-                    )
-
-                break
-
-        if component.type == StackComponentType.EXPERIMENT_TRACKER:
-            experiment_trackers.append(component.name)
-        if component.type == StackComponentType.STEP_OPERATOR:
-            step_operators.append(component.name)
-
-    settings_model = create_model("Settings", **settings_fields)
-
-    generic_step_fields = {}
-
-    for key, field_info in StepConfigurationUpdate.model_fields.items():
-        if key in [
-            "name",
-            "outputs",
-            "step_operator",
-            "experiment_tracker",
-            "parameters",
-        ]:
-            continue
-
-        generic_step_fields[key] = (field_info.annotation, field_info)
-
-    if experiment_trackers:
-        experiment_tracker_enum = Enum(
-            "ExperimentTrackers", {e: e for e in experiment_trackers}
-        )
-        generic_step_fields["experiment_tracker"] = (
-            experiment_tracker_enum,
-            None,
-        )
-    if step_operators:
-        step_operator_enum = Enum(
-            "StepOperators", {s: s for s in step_operators}
-        )
-        generic_step_fields["step_operator"] = (step_operator_enum, None)
-
-    generic_step_fields["settings"] = (settings_model, None)
-
-    all_steps = {}
-    for name, step_config in config.steps.items():
-        step_fields = generic_step_fields.copy()
-        if step_config.parameters:
-            parameter_fields = {
-                name: (Any, None) for name in step_config.parameters
-            }
-            parameters_class = create_model(
-                f"{name}_parameters", **parameter_fields
-            )
-            step_fields["parameters"] = (parameters_class, None)
-
-        step_model = create_model(name, **step_fields)
-        all_steps[name] = (step_model, None)
-
-    all_steps_model = create_model("Steps", **all_steps)
-
-    top_level_fields = {}
-
-    for key, field_info in PipelineRunConfiguration.model_fields.items():
-        if key in ["schedule", "build", "steps", "settings", "parameters"]:
-            continue
-
-        top_level_fields[key] = (field_info.annotation, field_info)
-
-    top_level_fields["settings"] = (settings_model, None)
-    top_level_fields["steps"] = (all_steps_model, None)
-
-    return create_model("Result", **top_level_fields).model_json_schema()
